1) hybrid Var-Ensemble schemes, which enable the the BE can be flow dependent. Contributed by Dale Baker and Yongsheng Chen. 2) Bug fix for compilation of makedepf90.
All the changes are limited in Var directory.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3347 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ var/scripts/da_perturb_wrf_bc.ksh
@@ -32,11 +32,9 @@ export RC_DIR_SAVE=$RC_DIR
 export NL_ANALYSIS_TYPE_SAVE=$NL_ANALYSIS_TYPE
 export CYCLING_SAVE=$CYCLING
 export RUN_DIR_SAVE=$RUN_DIR
- 
+
 #These are the local values:
-export END_DATE=$($BUILD_DIR/da_advance_time.exe $DATE $LBC_FREQ 2>/dev/null)
-export RC_DIR=$RUN_DIR_SAVE/rc
-mkdir -p $RC_DIR
+export END_DATE=$($BUILD_DIR/da_advance_time.exe $DATE $CYCLE_PERIOD 2>/dev/null)
 
 export NL_ANALYSIS_TYPE="randomcv"
 export NL_PUT_RAND_SEED=.TRUE.
@@ -46,21 +44,22 @@ if [[ $MEM -lt 100 ]]; then export CMEM=e0$MEM; fi
 if [[ $MEM -lt 10 ]]; then export CMEM=e00$MEM; fi
 
 while [[ $DATE -le $END_DATE ]]; do 
+
    export RUN_DIR=$RUN_DIR_SAVE/run/$DATE_SAVE/wrfvar/${DATE}.${CMEM}
    mkdir -p $RUN_DIR
 
-   export YEAR=$(echo $DATE | cut -c1-4)
-   export MONTH=$(echo $DATE | cut -c5-6)
-   export DAY=$(echo $DATE | cut -c7-8)
-   export HOUR=$(echo $DATE | cut -c9-10)
-   export ANALYSIS_DATE=${YEAR}-${MONTH}-${DAY}_${HOUR}:00:00
-   export NL_SEED_ARRAY1=$DATE_SAVE
-   export NL_SEED_ARRAY2=$(expr $DATE + $MEM)
+   export ANALYSIS_DATE=$($BUILD_DIR/da_advance_time.exe $DATE 0 -W 2>/dev/null)
+
+   export NL_SEED_ARRAY1=$($BUILD_DIR/da_advance_time.exe $DATE 0 -f hhddmmyycc)
+   export NL_SEED_ARRAY2=`echo $MEM \* 100000 | bc -l `
 
    echo "   Run WRF-Var in randomcv mode for date $DATE"
-   export DA_FIRST_GUESS=${RC_DIR}/$DATE/wrfinput_d${DOMAIN}
-   export DA_ANALYSIS=${RC_DIR}/$DATE/wrfinput_d${DOMAIN}.${CMEM}
-#   $SCRIPTS_DIR/da_trace.ksh da_run_wrfvar $RUN_DIR >&! /dev/null
+   export DA_FIRST_GUESS=${RC_DIR}/$DATE_SAVE/wrfinput_d01.${ANALYSIS_DATE}
+   export DA_ANALYSIS=${RC_DIR}/$DATE_SAVE/wrfinput_d01.${DATE}.${CMEM}
+#cys: need to set NL_ENSDIM_ALPHA=0, so that da_wrfvar.exe can run randomcv correctly
+   export NL_ENSDIM_ALPHA=0
+ 
+   $SCRIPTS_DIR/da_trace.ksh da_run_wrfvar $RUN_DIR
    $SCRIPTS_DIR/da_run_wrfvar.ksh > $RUN_DIR/index.html 2>&1
 
    RC=$?
@@ -71,23 +70,32 @@ while [[ $DATE -le $END_DATE ]]; do
 
    export NEXT_DATE=$($BUILD_DIR/da_advance_time.exe $DATE $LBC_FREQ 2>/dev/null)
    export DATE=$NEXT_DATE
+
 done
+
 
 echo "   Run pert_wrf_bc to create perturbed wrfbdy file for member $MEM"
 export RUN_DIR=$RUN_DIR_SAVE/run/$DATE_SAVE/pert_wrf_bc/${DATE_SAVE}.${CMEM}
 mkdir -p $RUN_DIR
 cd $RUN_DIR
 
-cp ${RC_DIR_SAVE}/${DATE_SAVE}/wrfbdy_d01 wrfbdy_this
-ln -fs ${RC_DIR}/${DATE_SAVE}/wrfinput_d01.${CMEM} wrfinput_this
-ln -fs ${RC_DIR}/${END_DATE}/wrfinput_d01.${CMEM} wrfinput_next
-ln -fs ${WPB_DIR}/input.nml .
-ln -fs ${WPB_DIR}/namelist.input .
-ln -fs ${WPB_DIR}/pert_wrf_bc.mac pert_wrf_bc.exe
-./pert_wrf_bc.exe > pert_wrf_bc.out.${CMEM} 2>&1
+export DATE=$DATE_SAVE
+while [[ $DATE -lt $END_DATE ]]; do
 
+   export NEXT_DATE=$($BUILD_DIR/da_advance_time.exe $DATE $LBC_FREQ 2>/dev/null)
+   if [[ ! -f wrfbdy_this ]]; then
+      cp ${RC_DIR_SAVE}/${DATE_SAVE}/wrfbdy_d01 wrfbdy_this
+   fi
+   ln -fs ${RC_DIR}/${DATE_SAVE}/wrfinput_d01.${DATE}.${CMEM} wrfinput_this
+   ln -fs ${RC_DIR}/${DATE_SAVE}/wrfinput_d01.${NEXT_DATE}.${CMEM} wrfinput_next
+   ln -fs ${WPB_DIR}/input.nml .
+   ln -fs ${WPB_DIR}/pert_wrf_bc pert_wrf_bc.exe
+   . $WRFVAR_DIR/inc/namelist_script.inc 
+   ./pert_wrf_bc.exe > pert_wrf_bc.out.${CMEM} 2>&1
+   export DATE=$NEXT_DATE
+done
 mv wrfbdy_this ${RC_DIR_SAVE}/${DATE_SAVE}/wrfbdy_d01.${CMEM}
-mv ${RC_DIR}/${DATE_SAVE}/wrfinput_d01.${CMEM} ${RC_DIR_SAVE}/${DATE_SAVE}
+#mv ${RC_DIR}/${DATE_SAVE}/wrfinput_d01.${CMEM} ${RC_DIR_SAVE}/${DATE_SAVE}
 
 export END_DATE=$END_DATE_SAVE
 export RC_DIR=$RC_DIR_SAVE
